Go back to using StreamableArrayList here, for now.
Events originating in Actionscript need to use that class because
we have no generic List class in actionscript. For things that
originate on the server this isn't a problem:
java array -> Array, and java List -> Array 
But for things originating on the actionscript client, any array
will become a java array unless we use StreamableArrayList.

So: maybe it's time to add List to the aspirin library. :|


git-svn-id: svn+ssh://src.earth.threerings.net/narya/trunk@6354 542714f4-19e9-0310-aa3c-eee0fc999fb1