First draft of mapping API. Mappings are a new way for sharing data between peers that provide more fine grained control over how data is distributed than just jamming your data into the NodeObject.
There are two kinds of mappings: mirrored and non-mirrored.

Mirrored mappings are (eventually) rebroadcast to all peers just as data in the
NodeObjects are, however one can choose to batch changes to those mappings and
only distribute them to peers periodically, so as to cut down on inter-peer
network traffic.

Non-mirrored mappings are maintained only on the mapping servers, and an
arbitrary peer must package up a unit of code and ship it to the mapping server
when they wish to read a mapping or process mappings in bulk. A mapping can be
added or removed from a non-mirrored mapping with a simple call. Non-mirrored
mappings substantially reduce inter-peer network traffic.