Allow downstream and upstream messages to be compounded
A client can collect a bunch of messages into a CompoundUpstreamMessage, which are unpacked and
dispatched in order by the server. The server collects any messages sent while processing that
compound message into a CompoundDownstreamMessage, which is unpacked and dispatched by the client.

This allows the client to indicate that it's going to send enough messages that it'd blow out the
throttle, but that it's doing so intentionally.



git-svn-id: svn+ssh://src.earth.threerings.net/narya/trunk@6659 542714f4-19e9-0310-aa3c-eee0fc999fb1