Yay, flexmojos 4.2-beta no longer uses wonky <scope>external</scope>.
It just leaves dependencies external by default when compiling an SWC, like any
sensible person would.

Also disabled warnings, since we never enabled them in the Ant build and
they're a giant list of spamming "hey, this class doesn't define a constructor"
warnings. Since when was it dangerous to rely on the default constructor?

git-svn-id: svn+ssh://src.earth.threerings.net/narya/trunk@6787 542714f4-19e9-0310-aa3c-eee0fc999fb1