Disable "manager calls" by default. **Breaking change**
This closes a giant security hole, but may break existing code.

PlaceObject.ManagerCaller is a poor-man's InvocationService, allowing
one to easily make a call to a method on the PlaceManager. The problem
is that this is turned on by default and any matching method is run.

This change disables it by default, but you may re-enable it by
overriding PlaceManager.handleManagerCalls() and returning true.

I also went ahead and made the ManagerCaller.invoke() method deprecated
so that you'll see a nice warning during compile time. You should
really consider just using a standard InvocationService to communicate with
the server. That way you effectively create a whitelist of client-callable
methods and there will be no surprises.